% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitt_palette.R
\docType{data}
\name{pitt_palettes}
\alias{pitt_palettes}
\title{List of color palettes.}
\format{
An object of class \code{list} of length 7.
}
\usage{
pitt_palettes
}
\description{
Use \code{\link{render_pitt_palette}} to construct palettes of desired length.
}
\keyword{datasets}
