% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitt_palette.R
\name{render_pitt_palette}
\alias{render_pitt_palette}
\title{Pitt palette rendering}
\usage{
render_pitt_palette(name, n)
}
\arguments{
\item{name}{Name of color palette. Choices are:
\code{"primary_contemporary"}, \code{"primary_legacy"}, \code{"primary"},
\code{"secondary"},
\code{"accent_rich"}, \code{"accent_vibrant"}, and \code{"accent"}.}

\item{n}{Number of colors desired. Most palettes have 3-6 colors. Most color
schemes are derived from \href{https://www.brand.pitt.edu/visual-identity/color-palettes-0}.
If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
Pitt palette rendering
}
\examples{
render_carnegie_colors("classic")
render_carnegie_colors("secondary", 5)
}
\keyword{colors}
